requirements.txt
Django==1.11.3
djangorestframework==3.6.3
Pygments==2.2.0
pytz==2017.2
